+++ src/server/api/routers/project.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 
 import { createTRPCRouter, protectedProcedure } from "~/server/api/trpc";
+import { TRPCError } from "@trpc/server";
 
 export const projectRouter = createTRPCRouter({
   getAll: protectedProcedure.query(({ ctx }) => {
@@ -15,6 +16,7 @@ export const projectRouter = createTRPCRouter({
     return ctx.prisma.project.findFirst({
       where: {
         id: input,
+        userId: ctx.session.user.id,
       },
     });
   }),
@@ -32,20 +34,55 @@ export const projectRouter = createTRPCRouter({
 
   update: protectedProcedure
     .input(z.object({ name: z.string(), id: z.string() }))
-    .mutation(({ ctx, input }) => {
-      return ctx.prisma.project.update({
+    .mutation(async ({ ctx, input }) => {
+      const project = await ctx.prisma.project.findUnique({
+        where: {
+          id: input.id,
+        },
+      });
+      if (!project) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Project does not exist",
+        });
+      }
+      if (project.userId !== ctx.session.user.id) {
+        throw new TRPCError({
+          code: "FORBIDDEN",
+          message: "You are not uthorized to update this project",
+        });
+      }
+      return ctx.prisma.project.updateMany({
         data: {
           name: input.name,
         },
         where: {
           id: input.id,
+          userId: ctx.session.user.id,
         },
       });
     }),
 
   delete: protectedProcedure
     .input(z.object({ id: z.string() }))
-    .mutation(({ ctx, input }) => {
+    .mutation(async ({ ctx, input }) => {
+      const project = await ctx.prisma.project.findUnique({
+        where: {
+          id: input.id,
+        },
+      });
+      if (!project) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Project does not exist",
+        });
+      }
+      if (project.userId !== ctx.session.user.id) {
+        throw new TRPCError({
+          code: "FORBIDDEN",
+          message: "You are not uthorized to delete this project",
+        });
+      }
       return ctx.prisma.project.delete({
         where: {
           id: input.id,
